Bump arrow to 55 (#1038)
Closes https://github.com/geoarrow/geoarrow-rs/issues/1037


### Change list

- Bump `arrow` to `55` and `parquet` to `55`
- Temporarily deactivates the `datafusion` integration until datafusion
publishes its version `47`
(https://github.com/apache/datafusion/issues/15072), so that we can
progress with the `arrow` 55 upgrade now.
- Update JS and Python APIs for latest `parquet`. 
- Means we no longer need an initial `HEAD` request for Parquet files
before reading metadata.

+++ rust/geoarrow/src/io/flatgeobuf/reader/object_store_reader.rs
@@ -9,7 +9,7 @@ use object_store::path::Path;
 pub struct ObjectStoreWrapper {
     pub location: Path,
     pub reader: Arc<dyn ObjectStore>,
-    pub size: usize,
+    pub size: u64,
 }
 
 #[async_trait]
@@ -19,10 +19,10 @@ impl AsyncHttpRangeClient for ObjectStoreWrapper {
         assert!(range.starts_with("bytes="));
 
         let split_range = range[6..].split('-').collect::<Vec<_>>();
-        let start_range = split_range[0].parse::<usize>().unwrap();
+        let start_range = split_range[0].parse::<u64>().unwrap();
 
         // Add one to the range because HTTP range strings are end-inclusive (I think)
-        let end_range = split_range[1].parse::<usize>().unwrap() + 1;
+        let end_range = split_range[1].parse::<u64>().unwrap() + 1;
 
         // Flatgeobuf will sometimes overfetch, but not all object store backends support
         // overfetches (e.g. this errors on a LocalFileSystem)
